InternalSensorManager jmri.managers -> jmri.jmrix.internal
