+++ java/src/jmri/configurexml/ClassMigration.properties
@@ -32,3 +32,5 @@ jmri.jmrit.display.configurexml.ControlPanelEditorXml  = jmri.jmrit.display.cont
 jmri.jmrit.display.configurexml.PanelEditorXml         = jmri.jmrit.display.panelEditor.configurexml.PanelEditorXml
 
 jmri.jmrit.signalling.configurexml.EntryExitPairsXml = jmri.jmrit.entryexit.configurexml.EntryExitPairsXml
+
+jmri.managers.configurexml.InternalSensorManagerXml = jmri.jmrix.internal.configurexml.InternalSensorManagerXml
