+++ java/src/jmri/jmrit/beantable/ReporterTableTabAction.java
@@ -12,21 +12,25 @@ public class ReporterTableTabAction extends AbstractTableTabAction<Reporter> {
         this("Multiple Tabbed");
     }
 
+    /** {@inheritDoc} */
     @Override
     protected Manager<Reporter> getManager() {
-        return InstanceManager.reporterManagerInstance();
+        return InstanceManager.getDefault(ReporterManager.class);
     }
 
+    /** {@inheritDoc} */
     @Override
     protected String getClassName() {
         return ReporterTableAction.class.getName();
     }
 
+    /** {@inheritDoc} */
     @Override
     protected ReporterTableAction getNewTableAction(String choice) {
         return new ReporterTableAction(choice);
     }
 
+    /** {@inheritDoc} */
     @Override
     protected String helpTarget() {
         return "package.jmri.jmrit.beantable.ReporterTable";
