+++ java/src/jmri/jmrit/beantable/SensorTableAction.java
@@ -411,7 +411,7 @@ public class SensorTableAction extends AbstractTableAction<Sensor> {
     protected void setDefaultState(JFrame _who) {
         String[] sensorStates = new String[]{Bundle.getMessage("BeanStateUnknown"), Bundle.getMessage("SensorStateInactive"), Bundle.getMessage("SensorStateActive"), Bundle.getMessage("BeanStateInconsistent")};
         JComboBox<String> stateCombo = new JComboBox<>(sensorStates);
-        switch (jmri.managers.InternalSensorManager.getDefaultStateForNewSensors()) {
+        switch (jmri.jmrix.internal.InternalSensorManager.getDefaultStateForNewSensors()) {
             case jmri.Sensor.ACTIVE:
                 stateCombo.setSelectedItem(Bundle.getMessage("SensorStateActive"));
                 break;
@@ -441,7 +441,7 @@ public class SensorTableAction extends AbstractTableAction<Sensor> {
             defaultState = jmri.Sensor.INCONSISTENT;
         }
 
-        jmri.managers.InternalSensorManager.setDefaultStateForNewSensors(defaultState);
+        jmri.jmrix.internal.InternalSensorManager.setDefaultStateForNewSensors(defaultState);
 
     }
