+++ java/src/jmri/jmrix/internal/configurexml/InternalSensorManagerXml.java
@@ -1,7 +1,9 @@
 package jmri.jmrix.internal.configurexml;
 
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import jmri.jmrix.internal.InternalSensorManager;
 import org.jdom2.Element;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 /**
  * Provides load and store functionality for configuring InternalSensorManagers.
@@ -11,8 +13,7 @@ import org.jdom2.Element;
  *
  * @author Bob Jacobsen Copyright: Copyright (c) 2006
  */
-@SuppressFBWarnings(value = "NM_SAME_SIMPLE_NAME_AS_SUPERCLASS", justification = "name assigned historically")
-public class InternalSensorManagerXml extends jmri.managers.configurexml.InternalSensorManagerXml {
+public class InternalSensorManagerXml extends jmri.managers.configurexml.AbstractSensorManagerConfigXML {
 
     public InternalSensorManagerXml() {
         super();
@@ -22,4 +23,56 @@ public class InternalSensorManagerXml extends jmri.managers.configurexml.Interna
     public void setStoreElementClass(Element sensors) {
         sensors.setAttribute("class", this.getClass().getName());
     }
+
+    @Override
+    public void load(Element element, Object o) {
+        log.error("Invalid method called");
+    }
+
+    @Override
+    public Element store(Object o) {
+        Element sensors = new Element("sensors");
+
+        String defaultState;
+        switch (InternalSensorManager.getDefaultStateForNewSensors()) {
+            case jmri.Sensor.ACTIVE:
+                defaultState = "active";
+                break;
+            case jmri.Sensor.INACTIVE:
+                defaultState = "inactive";
+                break;
+            case jmri.Sensor.INCONSISTENT:
+                defaultState = "inconsistent";
+                break;
+            default:
+                defaultState = "unknown";
+        }
+
+        sensors.addContent(new Element("defaultInitialState").addContent(defaultState));
+
+        return store(o, sensors);
+
+    }
+
+    @Override
+    public boolean load(Element sharedSensors, Element perNodeSensors) throws jmri.configurexml.JmriConfigureXmlException {
+        // load individual sharedSensors
+        if (sharedSensors.getChild("defaultInitialState") != null) {
+            int defaultState = jmri.Sensor.UNKNOWN;
+            String state = sharedSensors.getChild("defaultInitialState").getText();
+            if (state.equals("active")) {
+                defaultState = jmri.Sensor.ACTIVE;
+            } else if (state.equals("inactive")) {
+                defaultState = jmri.Sensor.INACTIVE;
+            } else if (state.equals("inconsistent")) {
+                defaultState = jmri.Sensor.INCONSISTENT;
+            }
+            InternalSensorManager.setDefaultStateForNewSensors(defaultState);
+        }
+        boolean load = loadSensors(sharedSensors);
+
+        return load;
+    }
+
+    private final static Logger log = LoggerFactory.getLogger(InternalSensorManagerXml.class);
 }
